Training that converges

+++ dataset.py
@@ -26,11 +26,11 @@ class CustomImageDataset(Dataset):
         if self.load_embeddings:
             emb_path = f"challenge/{labels_map[label]}/c{label+1}_{str(class_sample).zfill(3)}.pt"
             emb = torch.load(emb_path)
-            return emb, one_hot(torch.tensor(label), num_classes=4)
+            return emb, one_hot(torch.tensor(label), num_classes=4).to(torch.float)
         else:
             img_path = f"challenge/{labels_map[label]}/c{label+1}_{str(class_sample).zfill(3)}.tif"
             image = Image.open(img_path)
             if self.transform:
                 image = self.transform(image)
-            return image, one_hot(torch.tensor(label), num_classes=4)
+            return image, one_hot(torch.tensor(label), num_classes=4).to(torch.float)
